requirements.txt
playwright
playwright-stealth
requests
python-dotenv
google-genai
